requirements/base.txt
Django==1.10.8
django-rest-swagger==2.1.2
djangorestframework==3.7.1
djangorestframework-jwt==1.11.0
MySQL-python==1.2.5
celery==4.0.2
uWSGI==2.0.15
supervisor==3.3.2
django-cors-headers==1.3.1
django-redis==4.8.0
voluptuous==0.9.3
shortuuid==0.5.0
Pillow==5.2.0
django-filter==1.0.2